개인 회원가입 유효성 문제 -중복확인 -이메일 -생일 -동의(동의 없어도 연동됨) -가입 됬을때 가입 완료창 업체 회원가입 -우편 번호 클릭 안됨 - 500 error

+++ mybatis-spring-web/src/main/java/model/Member.java
@@ -12,9 +12,9 @@ public class Member {
    private String mAddr;
    private String mAddr_d;
    private String mPost;
-   private int mAttendcount;
+   private Long mAttendcount;
    private String mProfile;
-   private int mLicense;
+   private Long mLicense;
    private String mAuthority;
    private String bank;
    private String account;
@@ -32,7 +32,7 @@ public void setmPost(String mPost) {
    this.mPost = mPost;
 }
 public Member(String mEmail,String mPw,String mName,String mTel,
-      Date mBirth,String mGen,String mPost, String mAddr,String mAddr_d, int mAttendcount, String mAuthority, int mLicense
+      Date mBirth,String mGen,String mPost, String mAddr,String mAddr_d, Long mAttendcount, String mAuthority, Long mLicense
       ,String mProfile,String bank,String account) {
       this.mEmail = mEmail; 
       this.mPw = mPw;
@@ -110,11 +110,11 @@ public void setmAddr(String mAddr) {
    this.mAddr = mAddr;
 }
 
-public int getmAttendcount() {
+public Long getmAttendcount() {
    return mAttendcount;
 }
 
-public void setmAttendcount(int mAttendcount) {
+public void setmAttendcount(Long mAttendcount) {
    this.mAttendcount = mAttendcount;
 }
 
@@ -126,11 +126,11 @@ public void setmProfile(String mProfile) {
    this.mProfile = mProfile;
 }
 
-public int getmLicense() {
+public Long getmLicense() {
    return mLicense;
 }
 
-public void setmLicense(int mLicense) {
+public void setmLicense(Long mLicense) {
    this.mLicense = mLicense;
 }
